requirements.txt
Flask==3.0.3
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
Werkzeug==3.0.3
openpyxl==3.1.2
fpdf2==2.7.9